requirements.txt
Flask==3.1.3
openai-whisper==20250625
ffmpeg-python==0.2.0
torch==2.10.0
torchaudio==2.10.0
